Patch folly to use strong symbols
This is due to a recent change in folly, which defaulted to weak symbols
for jemalloc/tcmalloc. What this means is that they aren't linked into
folly and are instead loaded at runtime. The problem here is that Cling
fails to find them when JIT compiling, which then causes issues.
Manually linking to jemalloc or tcmalloc in jank doesn't fix the issue;
it leads to crashes instead. However, just having folly link in jemalloc
has been working, up until that recent change, so this patch just keeps
that going.